06/08/25 - move queries to module
